+++ packages/facility-server/app/routes/apiv1/index.js
@@ -6,7 +6,6 @@ import asyncHandler from 'express-async-handler';
 import { keyBy, mapValues } from 'lodash';
 
 import { allergy } from './allergy';
-import { addressHierarchy } from './addressHierarchy';
 import { appointments } from './appointments';
 import { asset } from './asset';
 import { attachment } from './attachment';
@@ -151,7 +150,6 @@ referenceDataRoutes.use('/user', user);
 referenceDataRoutes.use('/template', template);
 referenceDataRoutes.use('/vaccinationSettings', vaccinationSettings);
 referenceDataRoutes.use('/translation', translation);
-referenceDataRoutes.use('/addressHierarchy', addressHierarchy);
 
 // sync endpoints
 syncRoutes.use('/sync', sync);
