+++ packages/web/app/api/queries/index.js
@@ -1,4 +1,4 @@
-export * from './useAddressHierarchyQuery.js';
+export * from './useAddressHierarchyQuery';
 export * from './useVitalsSurveyQuery';
 export * from './useReferenceData';
 export * from './usePatientAdditionalDataQuery';
