+++ packages/web/app/api/queries/useAddressHierarchyQuery.js
@@ -4,7 +4,7 @@ import { useApi } from '../useApi';
 export const useAddressHierarchyQuery = queryParams => {
   const api = useApi();
 
-  return useQuery(['addressHierarchy', queryParams], () =>
-    api.get(`addressHierarchy`, queryParams),
+  return useQuery(['addressHierarchyTypes', queryParams], () =>
+    api.get('referenceData/addressHierarchyTypes', queryParams),
   );
 };
